RT74765. twr-k60n512: expand RAM_BUF to put u-boot.bin images there
Self-upgrade is possible now:
	tftp 0x1fff5800 aspotashev/k60/u-boot.bin
	cptf 0 ${fileaddr} ${filesize} 1

Since there is no external RAM on the TWR-K60N512 board, we put
the new U-Boot image into SRAM. The address of the RAM_BUF region
(0x1fff5800 in the above example) can be checked in the `u-boot.map`
file.

The RAM_BUF region is currently 84 Kbytes in size. U-Boot images larger
than this size cannot be safely loaded.